fixed registerDefaultTransforms test
- extracted editor creation to test utils ready for re-use in individual transform tests
- transforms are not run when using `setEditorState()` so we need to register them after that as they _are_ run when registering
- we can't use `getEditorState()` immediately after registering transforms because they are still being run async, we can get around that by running an empty discrete update first
